Restore window points when buffer is updated
It's nasty when each window point moves to the end when a new frame is
received and the content of the buffer is reinserted. Maybe you want
to have two windows for the pp-client buffer:

  - a window for the top of the buffer to watch the votes
  - a window for the bottom of the buffer to watch the log messages

Then it's good to restore the window points and keep the end of buffer
points if present.